set default tldraw state in setup
