+++ src/utils/userSettings.ts
@@ -1,15 +1,26 @@
-import { TDSnapshot } from "@tldraw/tldraw";
+import { TDSnapshot, TldrawApp } from "@tldraw/tldraw";
 
-export const STORAGE_SETTINGS_KEY = "sc_tldraw_settings";
+const STORAGE_SETTINGS_KEY = "sc_tldraw_settings";
 
-export const getUserSettings = (): TDSnapshot["settings"] | undefined => {
+const getUserSettings = (): TDSnapshot["settings"] | undefined => {
   const settingsString = localStorage.getItem(STORAGE_SETTINGS_KEY);
   return settingsString ? JSON.parse(settingsString) : undefined;
 };
 
-export const getDarkModeSetting = (): boolean | undefined => {
+const getDarkModeSetting = (): boolean | undefined => {
   const settings = getUserSettings();
 
   // returning undefined means that the system default will be used
   return settings ? settings.isDarkMode : undefined;
 };
+
+const setDefaultState = () => {
+  const userSettings = getUserSettings();
+  if (userSettings) {
+    TldrawApp.defaultState.settings = userSettings;
+  } else {
+    TldrawApp.defaultState.settings.language = "de";
+  }
+};
+
+export { STORAGE_SETTINGS_KEY, setDefaultState, getDarkModeSetting };
